@@ -1,60 +1,109 @@
-[project]
-name = "jmopenorders"
-version = "0.2.0"
-description = "A BDE tool"
-authors = [
-    "Jürgen Mülbert <juergen.muelbert@gmail.com>"
-]
-license = "EUPL-1.2"
-readme = "README.rst"
-python = "^3.6"
-homepage = "https://jmopenorders.github.io/"
-repository = "https://github.com/jmuelbert/jmopenorders"
-documentation = "https://jmopenorders.readthedocs.io/en/latest/"
-
-keywords = ['reporting', 'excel']
-
-classifiers = [
-        "Development Status:: 4 - Beta",
-        "Environment :: Console",
-        "Intended Audience:: Developers",
-        "Intended Audience :: End Users/Desktop",
-        "License:: OSI Approved:: European Union Public Licence 1.2 (EUPL 1.2)",
-        "Natural Language:: English",
-        "Operating System:: OS Independent",
-        "Operating System:: MacOS:: MacOS X",
-        "Operating System:: POSIX",
-        "Operating System:: POSIX:: BSD",
-        "Operating System:: POSIX:: Linux",
-        "Operating System:: Microsoft:: Windows",
-        "Programming Language:: Python",
-        "Programming Language:: Python:: 3",
-        "Programming Language:: Python:: 3:: Only",
-        "Programming Language:: Python:: 3.6",
-        "Programming Language:: Python:: 3.7",
-        "Programming Language:: Python:: Implementation:: CPython",
-        "Programming Language :: Python :: Implementation :: PyPy",
-        "Environment:: Console",
-        "Topic:: Office / Business",
-]
-
-# Requirements
-[dependencies]
-openpyxl="^3.0.0"
-PyYaml="^5.3.1"
-pip="^19.3.1"
-
-[dev-dependencies]
-black = { version = "^19.3-alpha.0", python = "^3.6" }
+# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject
 
 [build-system]
-# Minimum requirements for the build system to execute.
-# See https://github.com/scipy/scipy/pull/10431 for the AIX issue.
-requires = [
-    "setuptools >= 41.4.0",
-    "wheel >= 0.33.6",
+requires = ["poetry>=0.12"]
+build-backend = "poetry.masonry.api"
+
+[tool.poetry]
+name = "jmopenorders"
+version = "0.2.1.dev0"
+description = "Jmopenorders"
+authors = ["Jürgen Mülbert <juergen.muelbert@gmail.com>"]
+license = "EUPL-1.2"
+readme = "README.rst"
+homepage = "https://github.com/jmuelbert/jmopenorders"
+repository = "https://github.com/jmuelbert/jmopenorders"
+documentation = "https://jmopenorders.readthedocs.io"
+
+# Keywords description https://python-poetry.org/docs/pyproject/#keywords
+keywords = []  # Update me
+
+# Pypi classifiers: https://pypi.org/classifiers/
+classifiers=[
+  "Development Status :: 4 - Beta",
+  "Environment :: Console",
+  "Intended Audience :: Developers",
+  "Intended Audience :: End Users/Desktop",
+  "License :: OSI Approved :: European Union Public Licence 1.2 (EUPL 1.2)",
+        "Natural Language :: English",
+        "Operating System :: OS Independent",
+        "Programming Language :: Python",
+        "Programming Language :: Python :: 3",
+        "Programming Language :: Python :: 3.6",
+        "Programming Language :: Python :: 3.7",
+        "Programming Language :: Python :: 3.8",
+        "Programming Language :: Python :: Implementation :: CPython",
+        "Programming Language :: Python :: Implementation :: PyPy",
+        "Topic :: Office/Business",
 ]
-build-backend = 'setuptools.build_meta'
+
+[tool.poetry.scripts]
+# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
+jmopenorders = "jmopenorders.cli:main"
+
+[tool.poetry.urls]
+Changelog = "https://github.com/jmuelbert/jmopenorders/releases"
+
+[tool.poetry.dependencies]
+python = "^3.7"
+importlib_metadata = {version = "^1.6.0", python = "<3.8"}
+typer = {extras = ["all"], version = "^0.2.1"}
+rich = "^1.3.0"
+openpyxl = "^3.0.3"
+
+[tool.poetry.dev-dependencies]
+bandit = "^1.6.2"
+black = {version = "^black 19.10b0",allow-prereleases = true}
+bumpversion = "^0.6.0"
+darglint = "^1.4.0"
+doc8 = "^0.8"
+flake8-awesome = "^1.2.1"
+isort = {extras = ["pyproject"], version = "^4.3.21"}
+m2r = "^0.2"
+mypy = "^0.780"
+pydocstyle = "^5.0.2"
+pylint = "^2.5.2"
+pytest = "^5.4.3"
+pytest-cov = "^2.9.0"
+pytest-deadfixtures = "^2.2.0"
+pytest-mock = "^3.1.1"
+pytest-randomly = "^3.4"
+pyupgrade = "^2.4.4"
+unify = "^0.5"
+coverage = {extras = ["toml"], version = "^5.1"}
+safety = "^1.9.0"
+sphinx = "^3.0.4"
+sphinx-autobuild = "^0.7.1"
+sphinx-autodoc-typehints = "^1.10"
+tomlkit = "^0.6"
+typeguard = "^2.8.0"
+xdoctest = "^0.12.0"
+pre-commit = "^2.4.0"
+faker = "^4.1.0"
 
 [tool.black]
+# https://github.com/psf/black
 line-length = 80
+target-version = ["py37"]
+
+[tool.isort]
+# https://github.com/timothycrosley/isort/
+known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
+sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
+include_trailing_comma = true
+default_section = "FIRSTPARTY"
+multi_line_output = 3
+force_grid_wrap = 0
+use_parentheses = true
+line_length = 80
+
+[tool.coverage.paths]
+source = ["src", "*/site-packages"]
+
+[tool.coverage.run]
+branch = true
+source = ["jmopenorders"]
+
+[tool.coverage.report]
+show_missing = true
+fail_under = 100
