@@ -15,13 +15,11 @@
 # add these directories to sys.path here. If the directory is relative to the
 # documentation root, use os.path.abspath to make it absolute, like shown here.
 #
+"""Config for the sphinx-docu."""
 import datetime as dt
 import inspect
 import os
 import sys
-
-# import sphinx_rtd_theme
-
 
 __version__ = "0.2.1.dev0"
 __short_version__ = __version__
@@ -43,7 +41,7 @@ PROJECT_GITHUB_REPOSITORY = "jmopenorders"
 GITHUB_PATH = "{}/{}".format(
     PROJECT_GITHUB_USERNAME, PROJECT_GITHUB_REPOSITORY,
 )
-GITHUB_URL = "https://github.com/{}".format(GITHUB_PATH)
+GITHUB_URL = f"https://github.com/{GITHUB_PATH}"
 
 sys.path.insert(0, os.path.abspath("_ext"))
 sys.path.insert(0, os.path.abspath("../jmopenorders"))
@@ -52,7 +50,6 @@ sys.path.insert(0, os.path.abspath("../jmopenorders"))
 
 # If your documentation needs a minimal Sphinx version, state it here.
 #
-# needs_sphinx = '1.0'
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
@@ -70,7 +67,6 @@ templates_path = ["_templates"]
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
-# source_suffix = ['.rst', '.md']
 source_suffix = ".rst"
 
 # The master toctree document.
@@ -78,7 +74,7 @@ master_doc = "index"
 
 # General information about the project.
 project = PROJECT_NAME
-copyright = PROJECT_COPYRIGHT
+project_copyright = PROJECT_COPYRIGHT
 author = PROJECT_AUTHOR
 
 # The version info for the project you're documenting, acts as replacement for
@@ -90,7 +86,10 @@ version = __short_version__
 # The full version, including alpha/beta/rc tags.
 release = __version__
 
-code_branch = "dev" if "dev" in __version__ else "master"
+if "dev" in __version__:
+    code_branch = "develop"
+else:
+    code_branch = "master"
 
 # Edit on Github config
 edit_on_github_project = GITHUB_PATH
@@ -133,7 +132,7 @@ def linkcode_resolve(domain, info):
 
     fn = fn[index:]
 
-    return "{}/blob/{}/{}{}".format(GITHUB_URL, code_branch, fn, linespec)
+    return f"{GITHUB_URL}/blob/{code_branch}/{fn}{linespec}"
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
@@ -171,7 +170,6 @@ html_theme = "alabaster"
 # documentation.
 #
 html_theme_options = {
-    # 'logo': 'logo.png',
     "logo_name": PROJECT_NAME,
     "description": PROJECT_LONG_DESCRIPTION,
     "github_user": PROJECT_GITHUB_USERNAME,
@@ -179,33 +177,26 @@ html_theme_options = {
     "github_type": "star",
     "github_banner": True,
     "travis_button": True,
-    # 'touch_icon': 'logo-apple.png',
-    # 'fixed_sidebar': True, # Re-enable when we have more content
 }
 
 # Add any paths that contain custom themes here, relative to this directory.
-# html_theme_path = []
 
 # The name for this set of Sphinx documents.
 # "<project> v<release> documentation" by default.
 #
-# html_title = 'Home-Assistant v0.27.0'
 
 # A shorter title for the navigation bar.  Default is the same as html_title.
 #
-# html_short_title = None
 
 # The name of an image file (relative to this directory) to place at the top
 # of the sidebar.
 #
-# html_logo = '_static/logo.png'
 
 # The name of an image file (relative to this directory) to use as a favicon of
 # the docs.
 # This file should be a Windows icon file (.ico) being 16x16 or 32x32
 # pixels large.
 #
-# html_favicon = '_static/favicon.ico'
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -233,7 +224,6 @@ html_sidebars = {
 # .htaccess) here, relative to this directory. These files are copied
 # directly to the root of the documentation.
 #
-# html_extra_path = []
 
 # If not None, a 'Last updated on:' timestamp is inserted at every page
 # bottom, using the given strftime format.
@@ -257,16 +247,16 @@ htmlhelp_basename = "jmbdedoc"
 latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
-    # 'papersize': 'letterpaper',
-    # The font size ('10pt', '11pt' or '12pt').
+    # _papersize_: _letterpaper_,
+    # The font size _'10pt', '11pt' or '12pt'_.
     #
-    # 'pointsize': '10pt',
+    # _pointsize_: _10pt_,
     # Additional stuff for the LaTeX preamble.
     #
-    # 'preamble': '',
+    # _preamble_: __,
     # Latex figure (float) alignment
     #
-    # 'figure_align': 'htbp',
+    # _figure_align_: _htbp_,
 }
 
 # Grouping the document tree into LaTeX files. List of tuples
@@ -288,8 +278,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    (master_doc, "jmopenorders.tex",
-     "jmopenorders Documentation", [author], 1,)
+    (master_doc, "jmopenorders.tex", "jmopenorders Documentation", [author], 1,)
 ]
 
 
