@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2019 Jürgen Mülbert. All rights reserved.
+# Copyright (c) 2019-2020 Jürgen Mülbert. All rights reserved.
 #
 # Licensed under the EUPL, Version 1.2 or – as soon they
 # will be approved by the European Commission - subsequent
@@ -35,7 +35,7 @@
 # Die sprachspezifischen Genehmigungen und Beschränkungen
 # unter der Lizenz sind dem Lizenztext zu entnehmen.
 #
-""" Current version of the jmopenorders application.
+"""Current version of the jmopenorders application.
 
 This project uses the Semantic Versioning scheme in conjunction with PEP 0440:
 
@@ -50,6 +50,6 @@ incomplete states of a release .
 
 Version 0.x should be considered a development version with an unstable API,
 and backwards compatibility is not guaranteed for minor versions.
-
 """
-__version__ = "0.2.0"
+
+__version__ = "0.2.1.dev0"
